JWT 검증 적용

+++ controllers/likesController.js
@@ -1,38 +1,44 @@
 const conn = require("../mariadb");
 const { StatusCodes } = require("http-status-codes");
+const ensureAuthorization = require("../middlewares/ensureAuthorization");
+const checkTokenError = require("../middlewares/checkToeknError");
 
 const addLike = (req, res) => {
-  const { user_id } = req.body;
-  const { id } = req.params;
+  const bookId = req.params.id;
+  const auth = ensureAuthorization(req);
 
-  const addLikeQuery =
-    "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)";
+  if (!checkTokenError(auth, res)) {
+    const addLikeQuery =
+      "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)";
 
-  conn.query(addLikeQuery, [user_id, id], (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
+    conn.query(addLikeQuery, [auth.id, bookId], (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
 
-    return res.status(StatusCodes.OK).json(results);
-  });
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
 
 const cancelLike = (req, res) => {
-  const { user_id } = req.body;
-  const { id } = req.params;
+  const bookId = req.params.id;
+  const auth = ensureAuthorization(req);
 
-  const cancelLikeQuery =
-    "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?";
+  if (!checkTokenError(auth, res)) {
+    const cancelLikeQuery =
+      "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?";
 
-  conn.query(cancelLikeQuery, [user_id, id], (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
+    conn.query(cancelLikeQuery, [auth.id, bookId], (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
 
-    return res.status(StatusCodes.OK).json(results);
-  });
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
 
 module.exports = {
